Migrate rest of team endpoints to new pattern
